Add Creatures
Adds new tile type "Creature" denoted in maps by an "x" which
run around through empty space and generally cause havoc. Right now they
don't cause any direct damage to the player, though they make it really
easy to get stuck in spots. If we decide this is too annoying, we can
add Player to the list of tiles that these creatures bounce off of --
that'll make it so the player doesn't get stuck as much. Also adds
a map, "02", which highlights these new characters. It's probably too
chaotic but kind of fun for now. It's loosely based off one I saw in
a boulder dash YT video.